+++ src/main/java/com/arrowsmith/sakiladb/repositories/ActorRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ActorRepository extends JpaRepository<Actor, Short>
+public interface ActorRepository extends JpaRepository<Actor, Integer>
 {
 
 }
